#5043 Further changes to add event when sitemap created
