--- src/Libraries/Nop.Core/Domain/Common/SitemapCreatedEvent.cs
@@ -1,9 +1,0 @@
-﻿namespace Nop.Core.Domain.Common
-{
-    /// <summary>
-    /// Represents an event that occurs after the sitemap is created
-    /// </summary>
-    public class SitemapCreatedEvent
-    {
-    }
-}
