+++ src/Libraries/Nop.Services/Seo/ISitemapGenerator.cs
@@ -1,4 +1,4 @@
-using System.IO;
+﻿using System;
 
 namespace Nop.Services.Seo
 {
@@ -16,11 +16,15 @@ namespace Nop.Services.Seo
         string Generate(int? id);
 
         /// <summary>
-        /// This will build an XML sitemap for better index with search engines.
-        /// See http://en.wikipedia.org/wiki/Sitemaps for more information.
+        /// Get localized URLs
         /// </summary>
-        /// <param name="id">Sitemap identifier</param>
-        /// <param name="stream">Stream of sitemap.</param>
-        void Generate(Stream stream, int? id);
+        /// <param name="routeName">Route name</param>
+        /// <param name="routeParams">Lambda for route params object</param>
+        /// <param name="dateTimeUpdatedOn">A time when URL was updated last time</param>
+        /// <param name="updateFreq">How often to update url</param>
+        SitemapUrl GetLocalizedSitemapUrl(string routeName,
+            Func<int?, object> routeParams = null,
+            DateTime? dateTimeUpdatedOn = null,
+            UpdateFrequency updateFreq = UpdateFrequency.Weekly);
     }
 }
